Update the documentation.
It now includes the new `options` features (`options.ignore = {
$operatorToIgnore: true }` and `options.allow =
['$weirdOperatorToSupport']`).

It should be noted that the `options.allow` elements are trimmed by one
when looking for their correspondent methodNames. In other words,
operator names are always considered to be prefixed with '$' and have
methods with the same names as them, but without the prefix.